+++ examples/examples-component/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @engagespot/react-component@1.1.6
+
 ## 1.1.5
 
 ### Patch Changes